more work on wepon structure + beginning AI
fixed up the weapons to have their own indivdual triggers not all at the same time, created base stats for enemies and then children, one for ranged and one for melee. gave the melee logic to follow the player and attack when close and rigged up the damage for it. gave the player some set stats so they can take damage, die and began the ground work for mana / weapon dmg for player.